requirements.txt
requests
google-genai>=0.3.0
